requirements.txt
django-cors-headers==3.1.1
python-decouple==3.1
aniso8601==6.0.0
Babel==2.6.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Django==2.1.8
django-anymail==5.0
django-appconf==1.0.3
django-extensions==2.1.4
django-filter==2.0.0
django-graphql-extensions==0.0.6
django-graphql-jwt==0.2.0
django-imagekit==4.0.2
django-model-utils==3.1.2
django-mysql==2.4.1
django-queryset-csv==1.0.1
django-storages==1.7.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
factory-boy==2.12.0
Faker==2.0.2
google-api-core==1.14.0
google-api-python-client==1.7.5
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==0.28.1
google-cloud-storage==1.13.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
graphene==2.1.7
graphene-django==2.2.0
graphene-file-upload==1.2.2
graphene-permissions==1.1.2
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==19.9.0
httplib2==0.13.0
idna==2.8
inflection==0.3.1
oauth2client==4.1.3
olefile==0.46
pendulum==2.0.4
pilkit==2.0
Pillow==5.3.0
Pint==0.9
promise==2.2.1
protobuf==3.9.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycleanup==0.0.1
pyhumps==1.2.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.3.1
pypng==0.0.20
PyQRCode==1.2.1
python-dateutil==2.8.0
python-http-client==3.1.0
pytz==2018.9
pytzdata==2019.2
requests==2.21.0
rsa==4.0
Rx==1.6.1
sendgrid==6.0.2
singledispatch==3.4.0.3
six==1.12.0
stripe==2.17.0
text-unidecode==1.3
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.3
django-autoslug==1.9.6
xhtml2pdf==0.2.3
sqlparse==0.3.0

